refactor : entities 공통 모듈에서 도메인에 맞게 분리
