+++ src/collection/entities/collection-place.entity.ts
@@ -6,7 +6,7 @@ import {
   Column,
 } from 'typeorm';
 
-import { Place } from './place.entity';
+import { Place } from '../../place/entities/place.entity';
 import { Collection } from './collection.entity';
 
 @Entity()
