+++ src/place/entities/place-image.entity.ts
