+++ src/trip/entities/trip-style.entity.ts
